requirements.txt
requests
openai
pytest
allure-pytest
python-dotenv
pathspec>=0.11.1